Add files via upload
The file `app.py` is the main file where all the Flask code will be written. When you download this file, you also need to download `index.html` because it contains all the design elements, JavaScript functionalities, and styles. Additionally, you should download `background.jpg` for the background image. Ensure that you assemble these components in the correct order, and you'll be able to run the project on your desktop. This is a simple setup for the project.